fixed classes to be proper subclasses of ModelViewWithLighting so they have the getMCBoundingBoxes implemented and renders. Also added the other 3 decorative stairs to even the building out.

+++ FancyColumn.h
@@ -14,7 +14,7 @@ public:
 
 	cryph::AffPoint bottom;
 	cryph::AffPoint top;
-	float myBounds[6];
+	double myBounds[6];
 	float color[3];
 	float bradius;
 	float tradius;
